requirements.txt
numpy
pandas
yfinance
statsmodels
scipy
matplotlib
termcolor
alpaca-py
python-dotenv
streamlit
plotly
lxml
html5lib
